refactor: CreateMissedFile quickfix split in 2 classes
