+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkUnresolvedReferenceInspectionExtension.kt
@@ -1,9 +1,10 @@
 package com.jetbrains.snakecharm.inspections
 
 import com.intellij.codeInspection.LocalQuickFix
-import com.intellij.psi.*
+import com.intellij.psi.PsiReference
 import com.jetbrains.python.inspections.PyUnresolvedReferenceQuickFixProvider
-import com.jetbrains.snakecharm.inspections.quickfix.CreateMissedFile
+import com.jetbrains.snakecharm.inspections.quickfix.CreateMissedFileQuickFix
+import com.jetbrains.snakecharm.inspections.quickfix.CreateMissedFileUndoableAction
 import com.jetbrains.snakecharm.lang.psi.SmkArgsSection
 import com.jetbrains.snakecharm.lang.psi.SmkFileReference
 
@@ -12,13 +13,16 @@ class SmkUnresolvedReferenceInspectionExtension : PyUnresolvedReferenceQuickFixP
     override fun registerQuickFixes(reference: PsiReference, existing: MutableList<LocalQuickFix>) {
         val section = reference.element as? SmkArgsSection ?: return
         val sectionName = section.sectionKeyword ?: return
-        if (CreateMissedFile.supportedSections.containsKey(sectionName)) {
+        if (CreateMissedFileUndoableAction.sectionToDefaultFileContent.containsKey(sectionName)) {
             val fileReference = (section.reference as? SmkFileReference) ?: return
             if (!fileReference.hasAppropriateSuffix()) {
                 return
             }
             val name = fileReference.path
-            existing.add(CreateMissedFile(section, name, sectionName, fileReference.searchRelativelyToCurrentFolder))
+            existing.add(CreateMissedFileQuickFix(section,
+                name,
+                sectionName,
+                fileReference.searchRelativelyToCurrentFolder))
         }
     }
 }
